refactor: split display.ts into formatters, style, and renderer modules
Extract pure data-to-string helpers into shared/formatters.ts so both
the foreman and agent can import fmtEvent/fmtEventDetails/trunc/etc.
from a single canonical location — resolving the long-standing duplication
between display.ts and webhook-event.ts.

Extract terminal color/style constants (c, s, W, hr, effectiveWidth) into
src/agent/views/style.ts so renderer.ts can import them without creating
a circular dependency with display.ts.

Extract the markdown/diff rendering engine (renderMarkdown, renderTable,
fmtHunk, fmtBashOutput, fmtTodoWriteOutput, etc.) into
src/agent/views/renderer.ts as a self-contained content rendering module.

display.ts now contains only: Display class, clearBreak, format tables,
SDK block types, and the standalone resolve() helper. It re-exports c, s,
W, hr, and effectiveWidth from style.ts for backward compatibility.

webhook-event.ts drops ~50 lines of duplicated event-formatting helpers
and delegates to shared/formatters.ts instead.
