+++ src/agent/views/style.ts
@@ -1,0 +1,48 @@
+/**
+ * Terminal color, style, and layout constants.
+ * Extracted here so renderer.ts can import them without creating a
+ * circular dependency with display.ts.
+ */
+import { getConfig } from "../../config.js";
+
+// ── Display width ──────────────────────────────────────────────────────────
+
+export const W = 70;
+export const hr = (ch = "─") => ch.repeat(W);
+
+/** Visible width of the verbose timestamp prefix "HH:mm:ss " */
+const VERBOSE_PREFIX_LEN = 9;
+
+/**
+ * Returns the usable terminal width, accounting for the verbose timestamp
+ * prefix when verbose mode is active.
+ */
+export function effectiveWidth(fallback = W): number {
+  return (process.stdout.columns ?? fallback) - (getConfig().verbose ? VERBOSE_PREFIX_LEN : 0);
+}
+
+// ── Colors ─────────────────────────────────────────────────────────────────
+
+export const c = {
+  skyBlue:   (s: string) => `\x1b[38;5;117m${s}\x1b[0m`,
+  gray:      (s: string) => `\x1b[38;5;246m${s}\x1b[0m`,
+  amber:     (s: string) => `\x1b[38;5;214m${s}\x1b[0m`,
+  sageGreen: (s: string) => `\x1b[38;5;150m${s}\x1b[0m`,
+  salmon:    (s: string) => `\x1b[38;5;203m${s}\x1b[0m`,
+  boldRed:   (s: string) => `\x1b[1;31m${s}\x1b[0m`,
+  darkGray:  (s: string) => `\x1b[90m${s}\x1b[0m`,
+  yellow:    (s: string) => `\x1b[38;5;221m${s}\x1b[0m`,
+  lavender:  (s: string) => `\x1b[38;5;183m${s}\x1b[0m`,
+  bgGreen:   (s: string) => `\x1b[48;5;22m${s}\x1b[49m`,
+  bgRed:     (s: string) => `\x1b[48;5;52m${s}\x1b[49m`,
+};
+
+// ── Styles ─────────────────────────────────────────────────────────────────
+
+export const s = {
+  bold:          (s: string) => `\x1b[1m${s}\x1b[22m`,
+  dim:           (s: string) => `\x1b[2m${s}\x1b[22m`,
+  italic:        (s: string) => `\x1b[3m${s}\x1b[23m`,
+  underline:     (s: string) => `\x1b[4m${s}\x1b[24m`,
+  strikethrough: (s: string) => `\x1b[9m${s}\x1b[29m`,
+};
